requirements.txt
# Flask
Flask

# Flask extensions
Flask-Sqlalchemy
Flask-Wtf
Flask-Babel
Flask-Cache
Flask-Assets
Flask-Script
Flask-Bcrypt

# Misc 3rd party
slimish_jinja
augment
closure
jsmin
cssmin

# db migrations
alembic

# Dev requirements
flask-debugtoolbar
pylint
ipython
ipdb

gevent
